Zoom with magnifying gesture code refactored.
