+++ Muralisa/Muralisa/Views/SuggestionSubviews/ArtistSubview.swift
@@ -22,7 +22,7 @@ struct ArtistSubview: View {
 
     var body: some View {
         VStack(alignment: .leading) {
-            HStack {
+            VStack {
                 VStack {
                     switch manager.currentState {
                     case .loading:
@@ -65,6 +65,7 @@ struct ArtistSubview: View {
                                             .foregroundStyle(.gray)
                                             .opacity(0.2)
                                     }
+                                Spacer()
                             }
                             .padding(.bottom, 8)
                         }
@@ -73,13 +74,14 @@ struct ArtistSubview: View {
                         EmptyView()
                     }
                 }
-                Spacer()
                 
                 //Insertion date
                 HStack {
                     Image(systemName: "calendar")
                     Text(date)
+                    Spacer()
                 }
+                .padding(.vertical, 8)
                 .font(.footnote)
                 .foregroundStyle(.secondary)
             }
